first commit with build completions script deprecated and this step moved to completions install time
