+++ spatialprofilingtoolbox/entry_point/spt_enable_completion.py
@@ -4,6 +4,12 @@ from os.path import join
 from os.path import exists
 import argparse
 
+from jinja2 import Environment
+from jinja2 import BaseLoader
+
+from spatialprofilingtoolbox.entry_point.spt import get_commands
+from spatialprofilingtoolbox import submodule_names
+
 header = '### Start added by spatialprofilingtoolbox'
 footer = '### End added by spatialprofilingtoolbox'
 
@@ -28,6 +34,34 @@ def remove_previous_installation(filename):
         with open(filename, 'wt') as file:
             file.write(new_contents)
             print('Removed previous completion code from %s' % filename)
+
+
+def get_nontrivial_module_names():
+    return [name for name in submodule_names if len(get_commands(name)) > 0]
+
+
+def get_modules_and_commands():
+    return [
+        {
+            'name': module_name,
+            'command_names_joined_space': ' '.join(get_commands(module_name)),
+            'command_names_joined_bar': '|'.join(["'%s'" % c for c in get_commands(module_name)]),
+        }
+        for module_name in get_nontrivial_module_names()
+    ]
+
+
+def create_completions_script():
+    jinja_environment = Environment(
+        loader=BaseLoader, comment_start_string='###')
+    with importlib.resources.path('spatialprofilingtoolbox.entry_point',
+                                  'spt-completion.sh.jinja') as path:
+        with open(path, 'r') as file:
+            template_source = file.read().rstrip('\n')
+    template = jinja_environment.from_string(template_source)
+    modules = get_modules_and_commands()
+    return template.render(
+        module_names=' '.join(get_nontrivial_module_names()), modules=modules)
 
 
 def attempt_append_to(filename, contents):
@@ -64,12 +98,7 @@ def main_program():
                 remove_previous_installation(path)
         exit()
 
-    with importlib.resources.path('spatialprofilingtoolbox.entry_point',
-                                  'spt-completion.sh') as path:
-        with open(path, 'r') as file:
-            completion_script = file.read().rstrip('\n')
-    lines = completion_script.split('\n')
-    completion_script = '\n'.join(lines[1:])
+    completion_script = create_completions_script()
     wrapped = '\n%s\n%s\n%s\n' % (header, completion_script, footer)
 
     for file in profile_files:
